moved calculations into energyCalculations and moved main calculations into timer1 interrupt

+++ Basic/Jenny/209Project/209Project/uart.c
@@ -29,7 +29,7 @@ void trans_array(char* msg){
 	}
 }
 
-void trans_float(float value) {
+void trans_float(uint16_t value) {
 	char buffer[20];
 	snprintf(buffer, sizeof(buffer), "%f\n", value);
 	trans_array(buffer);
